fixed check number length

+++ src/main/kotlin/com/cynergisuite/middleware/accounting/account/payable/invoice/AccountPayableCheckPreviewVendorsDTO.kt
@@ -3,6 +3,7 @@ package com.cynergisuite.middleware.accounting.account.payable.invoice
 import com.fasterxml.jackson.annotation.JsonView
 import io.swagger.v3.oas.annotations.media.Schema
 import java.math.BigDecimal
+import java.math.BigInteger
 import java.time.LocalDate
 import javax.validation.constraints.NotNull
 
@@ -34,7 +35,7 @@ data class AccountPayableCheckPreviewVendorsDTO(
    var postalCode: String? = null,
 
    @field:Schema(description = "Check number")
-   var checkNumber: Int? = null,
+   var checkNumber: BigInteger? = null,
 
    @field:Schema(description = "Distribution amount")
    var date: LocalDate? = null,
