+++ src/main/kotlin/com/cynergisuite/middleware/accounting/account/payable/invoice/AccountPayableCheckPreviewVendorsEntity.kt
@@ -1,6 +1,7 @@
 package com.cynergisuite.middleware.accounting.account.payable.invoice
 
 import java.math.BigDecimal
+import java.math.BigInteger
 import java.time.LocalDate
 
 data class AccountPayableCheckPreviewVendorsEntity (
@@ -11,7 +12,7 @@ data class AccountPayableCheckPreviewVendorsEntity (
    val city: String?,
    val state: String?,
    val postalCode: String?,
-   val checkNumber: Int,
+   val checkNumber: BigInteger,
    val date: LocalDate,
    val invoiceList: MutableList<AccountPayableCheckPreviewInvoiceEntity>? = mutableListOf(),
    var gross: BigDecimal = BigDecimal.ZERO,
